refactor: 클래스 변경
WinningNumber를 삭제하고 Lotto의 형식에 맞춰 기능을 이동
