+++ docs/README.md
@@ -10,15 +10,15 @@
 
 **당첨 번호 생성**
 
-- [x] 6개의 당첨 번호를 입력받는다. - WinningNumber
+- [x] 6개의 당첨 번호를 입력받는다. - Lotto
     - [x] 당첨번호 입력값에서 공백을 모두 제거한다. - Util#removeSpace()
-    - [x] 쉼표를 기준으로 구분한다. - WinningNumber#separateStringByComma()
+    - [x] 쉼표를 기준으로 구분한다. - Lotto#separateStringByComma()
     - [x] 구분된 당첨번호의 개수가 6개가 아니면 예외 처리한다.
     - [x] 구분된 당첨번호가 1~45 사이의 숫자가 아니면 예외 처리한다. - Util#verifyRangeOfLottoNumber()
     - [x] 중복되는 숫자가 있으면 예외 처리한다.
 - [x] 보너스 번호 1개를 입력받는다. - BonusNumber
     - [x] 보너스 번호가 1~45 사이의 숫자가 아니면 예외 처리한다. - Util#verifyRangeOfLottoNumber()
-    - [x] 보너스 번호가 당첨번호와 중복되면 예외 처리한다. - BonusNumber#verifyDuplicateNumber()
+    - [ ] 보너스 번호가 당첨번호와 중복되면 예외 처리한다.
 
 **로또 발행**
