+++ src/main/java/lotto/model/BonusNumber.java
@@ -13,15 +13,15 @@ public class BonusNumber {
 
     private static int validateBonusNumber(int number) {
         verifyRangeOfLottoNumber(number);
-        verifyDuplicateNumber(number);
+        /*verifyDuplicateNumber(number);*/
         return number;
     }
 
-    private static void verifyDuplicateNumber(int number) {
-        if (WinningNumber.winningNumbers.contains(number)) {
+  /*  private static void verifyDuplicateNumber(int number) {
+        if (Lotto.getNumbers().contains(number)) {
             throw new IllegalArgumentException("[ERROR] 보너스 번호가 당첨 번호와 중복됩니다.");
         }
-    }
+    }*/
 
     private static int formatInput(String input) {
         return Integer.parseInt(removeSpace(input));
